refactor: move tool rendering logic into per-action files
- Add ProcessAction union type for type-safe action dispatch
- Extract formatTimestamp and formatStatusTag to utils/format.ts
- Add render exports to each action file (start, list, output, logs, kill, write, debug)
- Add renderActionCall and renderActionResult dispatchers to actions/index.ts
- Slim down tools/index.ts from 492 to 202 lines

Each action file now owns both execution and rendering logic for better co-location.
