+++ src/tools/actions/start.ts
@@ -1,5 +1,12 @@
-import type { ExtensionContext } from "@mariozechner/pi-coding-agent";
-import type { ExecuteResult } from "../../constants";
+import { ToolBody, ToolCallHeader } from "@aliou/pi-utils-ui";
+import type {
+  AgentToolResult,
+  ExtensionContext,
+  Theme,
+  ToolRenderResultOptions,
+} from "@mariozechner/pi-coding-agent";
+import { Text } from "@mariozechner/pi-tui";
+import type { ExecuteResult, ProcessesDetails } from "../../constants";
 import type { ProcessManager } from "../../manager";
 
 type WatchStream = "stdout" | "stderr" | "both";
@@ -17,6 +24,100 @@ interface StartParams {
   alertOnFailure?: boolean;
   alertOnKill?: boolean;
   logWatches?: StartLogWatch[];
+}
+
+export function renderStartCall(
+  args: StartParams,
+  theme: Theme,
+): ToolCallHeader {
+  const longArgs: Array<{ label?: string; value: string }> = [];
+  const optionArgs: Array<{ label: string; value: string }> = [];
+  let mainArg: string | undefined;
+
+  if (args.name) {
+    mainArg = `"${args.name}"`;
+  }
+
+  if (args.command) {
+    if (!mainArg && args.command.length <= 60) {
+      mainArg = args.command;
+    } else if (args.command.length <= 60) {
+      optionArgs.push({ label: "command", value: args.command });
+    } else {
+      longArgs.push({ label: "command", value: args.command });
+    }
+  }
+
+  if (args.logWatches && args.logWatches.length > 0) {
+    optionArgs.push({
+      label: "watches",
+      value: String(args.logWatches.length),
+    });
+  }
+
+  return new ToolCallHeader(
+    {
+      toolName: "Process",
+      action: "start",
+      mainArg,
+      optionArgs,
+      longArgs,
+    },
+    theme,
+  );
+}
+
+export function renderStartResult(
+  result: AgentToolResult<ProcessesDetails>,
+  options: ToolRenderResultOptions,
+  theme: Theme,
+): ToolBody {
+  const { details } = result;
+  const process = details.process;
+
+  if (!process) {
+    return new ToolBody(
+      {
+        fields: [
+          {
+            label: "Error",
+            value: "Missing process details",
+            showCollapsed: true,
+          },
+        ],
+      },
+      options,
+      theme,
+    );
+  }
+
+  const fields: Array<
+    { label: string; value: string; showCollapsed?: boolean } | Text
+  > = [
+    new Text(
+      [
+        theme.fg("success", "Started process"),
+        `  name: ${theme.fg("accent", process.name)}`,
+        `  command: ${process.command}`,
+        `  id: ${theme.fg("accent", process.id)}`,
+        `  pid: ${String(process.pid)}`,
+        "  Log files:",
+        `    - stdout: ${theme.fg("accent", process.stdoutFile)}`,
+        `    - stderr: ${theme.fg("accent", process.stderrFile)}`,
+      ].join("\n"),
+      0,
+      0,
+    ),
+    {
+      label: "Status",
+      value:
+        theme.fg("success", "Started") +
+        ` ${theme.fg("accent", `"${process.name}"`)} (${process.id}, PID: ${process.pid})`,
+      showCollapsed: true,
+    },
+  ];
+
+  return new ToolBody({ fields }, options, theme);
 }
 
 export function executeStart(
