+++ src/utils/format.ts
@@ -1,3 +1,4 @@
+import type { Theme } from "@mariozechner/pi-coding-agent";
 import type { ProcessInfo } from "../constants";
 
 export function formatRuntime(
@@ -39,4 +40,35 @@ export function formatStatus(proc: ProcessInfo): string {
 export function truncateCmd(cmd: string, max = 40): string {
   if (cmd.length <= max) return cmd;
   return `${cmd.slice(0, max - 3)}...`;
+}
+
+export function formatTimestamp(ts: number | null): string {
+  if (!ts) return "-";
+  return new Date(ts).toISOString().replace("T", " ").slice(0, 19);
+}
+
+export function formatStatusTag(
+  process: {
+    status: string;
+    success: boolean | null;
+    exitCode: number | null;
+  },
+  theme: Theme,
+): string {
+  switch (process.status) {
+    case "running":
+      return theme.fg("accent", "running");
+    case "terminating":
+      return theme.fg("warning", "terminating");
+    case "terminate_timeout":
+      return theme.fg("error", "terminate_timeout");
+    case "killed":
+      return theme.fg("warning", "killed");
+    case "exited":
+      return process.success
+        ? theme.fg("success", "exit(0)")
+        : theme.fg("error", `exit(${process.exitCode ?? "?"})`);
+    default:
+      return theme.fg("muted", process.status);
+  }
 }
