+++ src/utils/index.ts
@@ -1,3 +1,9 @@
 export { hasAnsi, stripAnsi } from "./ansi";
-export { formatRuntime, formatStatus, truncateCmd } from "./format";
+export {
+  formatRuntime,
+  formatStatus,
+  formatStatusTag,
+  formatTimestamp,
+  truncateCmd,
+} from "./format";
 export { isProcessGroupAlive, killProcessGroup } from "./process-group";
